requirements.txt
accelerate==1.7.0
ag2==0.9.1.post0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
asyncer==0.0.8
attrs==25.3.0
autogen==0.9.1.post0
av==14.4.0
bitsandbytes==0.41.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.1
datasets==2.16.1
decorator==5.2.1
decord==0.6.0
deepspeed==0.14.4
diffusers==0.33.1
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docker-pycreds==0.4.0
docstring_parser==0.16
-e git+ssh://git@github.com/mit-han-lab/efficientvit.git@71b35be1d45e569acf9e82eb8c55a47a7242cd1e#egg=efficientvit
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
flatbuffers==25.2.10
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2023.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gradio_client==0.2.9
h11==0.16.0
hf-xet==1.1.2
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
humanfriendly==10.0
idna==3.10
igraph==0.11.8
imageio==2.37.0
importlib_metadata==8.7.0
ipdb==0.13.13
ipython==8.36.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
latex2mathml==3.78.0
lazy_loader==0.4
lightning-utilities==0.14.3
-e git+ssh://git@github.com/LLaVA-VL/LLaVA-NeXT.git@b42941ceba259d5df18f8df8193a3897296a0449#egg=llava
lvis==0.5.3
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.15
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
onnxsim==0.4.36
open_clip_torch==2.32.0
openai==1.82.1
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
parso==0.8.4
peft==0.4.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyarrow-hotfix==0.7
pycocotools==2.0.8
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.11
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.11
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.2.2
scipy==1.15.3
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
sentencepiece==0.1.99
sentry-sdk==2.29.1
setproctitle==1.3.6
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
svgwrite==1.4.3
sympy==1.14.0
termcolor==3.1.0
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
timm==1.0.15
TinyNeuralNetwork @ git+https://github.com/alibaba/TinyNeuralNetwork.git@5efa231483aa5229f32cc125879f539f7e283c35
tokenizers==0.21.1
tomli==2.2.1
torch==2.1.2
torch-fidelity==0.3.0
torchmetrics==1.7.1
torchprofile==0.0.4
torchvision==0.16.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==2.1.0
typeguard==4.4.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tyro==0.9.22
tzdata==2025.2
ultralytics==8.3.147
ultralytics-thop==2.0.14
urllib3==1.26.20
uvicorn==0.34.2
wandb==0.19.11
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
